pyproject.toml
[project]
name = "PROD"
version = "0.1.0"
description = "PROD PromoV2 Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.14.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "async-timeout==5.0.1",
    "asyncpg==0.30.0",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "dishka==1.4.2",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.6",
    "fastapi-cli==0.0.7",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.5",
    "mako==1.3.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "passlib==1.7.4",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.37",
    "starlette==0.41.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "watchfiles==1.0.4",
    "websockets==14.1",
]

[tool.ruff]
line-length = 128
target-version = "py311"
extend-exclude = [
    ".git",
    ".dockerignore",
    ".gitignore",
    ".env",
    ".python-version",
    "alembic.ini",
    "Dockerfile",
    "pyproject.toml",
    "README.md",
    "requirements.txt",
    "ruff.toml",
    "uv.lock",
    "alembic",
    "tests",
]

[tool.ruff.lint]
extend-select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]
ignore = [
    "E501",
    "SIM102",
    "F821",
    "F401",
]

[tool.black]
line-length = 128
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.dockerignore
  | \.gitignore
  | \.env
  | \.python-version
  | alembic.ini
  | Dockerfile
  | pyproject.toml
  | README.md
  | requirements.txt
  | ruff.toml
  | uv.lock
  | alembic
  | tests
)/
'''
